Onboard workspace demonstrators-social/shared
No integration provided, for this reason the build gets broken
