+++ packages/demonstrators/social/shared/src/recoil/features/scroll/interaction/types.ts
@@ -1,0 +1,3 @@
+export interface InteractionOption {
+  activeIds: { [key: string]: boolean };
+}
