+++ packages/demonstrators/social/shared/src/recoil/features/scroll/post/types.ts
@@ -1,0 +1,10 @@
+import { EffectRef } from '@huse/effect-ref';
+
+export interface RefPostScroll {
+  refObserved: EffectRef<HTMLElement>;
+  refObservedSubSlices: { [key: string]: EffectRef<HTMLElement> };
+}
+
+export type RefPostScrollType = {
+  [value: string]: RefPostScroll;
+};
