+++ packages/demonstrators/social/shared/src/recoil/features/scroll/reducers.ts
@@ -1,0 +1,9 @@
+import { reducers as interaction } from './interaction/reducer';
+import { reducers as post } from './post/reducer';
+import { reducers as timeline } from './timeline/reducer';
+
+export const reducers = {
+  interaction,
+  post,
+  timeline
+};
