+++ packages/demonstrators/social/shared/src/recoil/features/scroll/selectors.ts
@@ -1,0 +1,7 @@
+import { selectors as interaction } from './interaction/reducer';
+import { selectors as post } from './post/reducer';
+
+export const selectors = {
+  interaction,
+  post
+};
