+++ packages/demonstrators/social/shared/src/recoil/features/scroll/states.ts
@@ -1,0 +1,9 @@
+import { states as interaction } from './interaction/reducer';
+import { states as post } from './post/reducer';
+import { states as timeline } from './timeline/reducer';
+
+export const states = {
+  interaction,
+  post,
+  timeline
+};
