+++ packages/demonstrators/social/shared/src/recoil/features/scroll/timeline/actions.ts
@@ -1,90 +1,9 @@
-import { EffectRef } from '@huse/effect-ref';
 import _ from 'lodash';
 import { Translate } from 'next-translate';
-import { atom, atomFamily } from 'recoil';
 
-import { Relation } from '../../../../components/archer/types';
-import { CollectionUtil } from '../../../../utils';
-
-export enum VIEW {
-  TIMELINE = 0,
-  POSTS = 1
-}
-
-export interface Views {
-  currentViews: { [key: string]: VIEW };
-}
-
-export const timelineViewState = atom<Views>({
-  key: 'timelineView',
-  default: {
-    currentViews: {}
-  }
-});
-
-export interface RefContainerScroll {
-  refObserved: EffectRef<HTMLElement>;
-}
-
-export const refContainerScrollState = atomFamily<RefContainerScroll, string>({
-  key: 'refContainerScroll',
-  default: {
-    refObserved: null
-  }
-});
-
-const updateObservedItem = (
-  state: RefContainerScroll,
-  value: EffectRef<HTMLElement>
-): RefContainerScroll => {
-  return {
-    ...state,
-    refObserved: value
-  };
-};
-
-const removeObservedItem = (state: RefContainerScroll): RefContainerScroll => {
-  const refObservedUpdated = CollectionUtil.Object.removePropertyFromObject(
-    state,
-    'refObserved'
-  );
-  return {
-    ...refObservedUpdated
-  };
-};
-
-export const scrollTimelineReducer = {
-  updateObservedItem,
-  removeObservedItem
-};
-
-export const refContainerScrollFromArcherState = atomFamily<
-  RefContainerScroll,
-  string
->({
-  key: 'refContainerScrollFromArcher',
-  default: {
-    refObserved: null
-  }
-});
-
-// Node and Relation
-// => publish (fixed)>-node-<(fixed) head
-// => publish (fixed)>-relation-<(fixed) head
+import { LAYOUT, Link, NodeWithRelationsWithEdge } from './types';
 
 export const publishActions = ['head', 'upload', 'download', 'tail'];
-export const publishActions2 = ['headB', 'uploadB', 'downloadB', 'tailB'];
-
-export const baseActions = ['head', 'tail'];
-
-type Link = {
-  id: string;
-  nodeValue: string;
-  value: string;
-};
-
-// => publish (fixed)>-node-<(fixed) head
-// => publish (fixed)>-relation-<(fixed) head
 
 export const addTopic = (
   items: Array<string>,
@@ -100,64 +19,6 @@ export const addTopic = (
     return acc;
   }, []);
 };
-
-export interface Node {
-  id: string;
-  label: string;
-}
-
-export interface NodeWithRelations {
-  relations: Array<Relation>;
-  node: Node;
-}
-
-export interface EdgeConnectionMap {
-  postIds: { [key: string]: string };
-  sliceIds: { [key: string]: string };
-  timelineIds: { [key: string]: string };
-}
-
-export enum LAYOUT {
-  FULL = 0,
-  PROGRESSIVE = 1
-}
-
-export interface NodeWithRelationsWithEdge {
-  nodeWithRelations: Array<NodeWithRelations>;
-  edgeConnections: EdgeConnectionMap;
-  ltr: boolean;
-  layout: LAYOUT;
-}
-
-// TODO: find ideal name
-export interface A {
-  values: Array<NodeWithRelationsWithEdge>;
-  id: string;
-  description: string;
-}
-
-export interface NodesWithRelationsMap {
-  nodesWithRelations: {
-    [key: string]: {
-      values: Array<NodeWithRelationsWithEdge>;
-      id: string;
-      description: string;
-    };
-  };
-  activeId: string;
-  counter: number;
-  finalSize: number;
-}
-
-export const nodesWithRelationsWithEdgeState = atom<NodesWithRelationsMap>({
-  key: 'nodesWithRelationsWithEdge',
-  default: {
-    nodesWithRelations: {},
-    activeId: '',
-    counter: 0,
-    finalSize: 0
-  }
-});
 
 const getEdgeConnections = (
   item: Link,
@@ -266,4 +127,13 @@ export const createNodesWithRelations = (
         layout
       });
   return result;
+};
+
+export const actions = {
+  addTopic,
+  createNodesWithRelations
+};
+
+export const data = {
+  publishActions
 };
