+++ packages/demonstrators/social/shared/src/redux/features/actionType.ts
@@ -1,0 +1,11 @@
+import { ActionType } from 'typesafe-actions';
+
+import * as timeline from './timeline/actions';
+
+export const actions = {
+  timeline
+};
+
+export type TimelineActions = ActionType<typeof timeline>;
+
+export type StoreAction = TimelineActions;
