+++ packages/demonstrators/social/shared/src/redux/features/selectors.ts
@@ -1,0 +1,5 @@
+import * as timeline from './timeline/selector';
+
+export const selectors = {
+  timeline
+};
