+++ packages/demonstrators/social/shared/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../tsconfig.settings.json",
+  "compilerOptions": {
+    "composite": true,
+    "rootDir": "src",
+    "outDir": "dist"
+  },
+  "include": ["**/*.ts"]
+}
